extract dataclasses for residences2018
